moved common services to shared folder
