+++ chatter/chat-app/src/app/shared/services/guards/auth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
-import { TokenService } from '../services/token.service';
+import { TokenService } from '../token.service';
 
 @Injectable({
   providedIn: 'root'
